feat(hooks): lift BiasedPotentialHook, NeighborListHook, WrapPeriodicHook to nvalchemi.hooks
Move three general-purpose hooks out of nvalchemi.dynamics.hooks into the
shared nvalchemi.hooks namespace so they can be used in non-dynamics
workflows without importing the dynamics sub-package.

- stage defaults to None (stage-agnostic); callers pass explicit stage
- wrap_positions custom op co-located in hooks/periodic.py
- old dynamics/hooks/{bias,neighbor_list,periodic}.py deleted (no shims)
